🐛 adjust broad/slender check for end mutations
The check for broad or slender ending was assuming that the broadness of the last vowel of the stem would align with the broadness for the ending. This was not necessarily the case if the second last consonant had a broad vowel on one side and a slender vowel on the other, e.g. "neamhnigh". Now the check is on full verb with different logic for different endings.